+++ NEWS.md
@@ -1,3 +1,5 @@
+# tern next release
+
 # tern 0.6.2
 
 * use chached data from `random.cdisc.data` to speed up testing